requirements.txt
Django>=5.0.1,<5.1.0
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary>=2.9.9,<3.0.0